+++ x/axelarcork/client/cli/query.go
@@ -10,17 +10,6 @@ import (
 	"github.com/spf13/cobra"
 )
 
-const (
-	FlagAxelarChainID   = "axelar-chain-id"
-	FlagAxelarChainName = "axelar-chain-name"
-)
-
-// AddChainFlagsToCmd adds common chain flags to a module command.
-func AddChainFlagsToCmd(cmd *cobra.Command) {
-	cmd.Flags().String(FlagAxelarChainName, "", "The case sensitive name of the Axelar target chain")
-	cmd.Flags().Uint64(FlagAxelarChainID, 0, "The Chain ID for the Axelar target EVM chain")
-}
-
 // GetQueryCmd returns the cli query commands for this module
 func GetQueryCmd() *cobra.Command {
 	corkQueryCmd := &cobra.Command{
@@ -89,21 +78,12 @@ func queryCellarIDs() *cobra.Command {
 			}
 
 			req := &types.QueryCellarIDsRequest{}
-			name, err := cmd.Flags().GetString(FlagAxelarChainName)
+			name, chainID, err := GetChainInfoFromFlags(cmd)
 			if err != nil {
 				return err
 			}
-			chainID, err := cmd.Flags().GetUint64(FlagAxelarChainID)
-			if err != nil {
-				return err
-			}
-
-			if name != "" {
-				req.ChainName = name
-			}
-			if chainID != 0 {
-				req.ChainId = chainID
-			}
+			req.ChainName = name
+			req.ChainId = chainID
 
 			queryClient := types.NewQueryClient(ctx)
 
@@ -137,21 +117,12 @@ func queryScheduledCorks() *cobra.Command {
 			queryClient := types.NewQueryClient(ctx)
 			req := &types.QueryScheduledCorksRequest{}
 
-			name, err := cmd.Flags().GetString(FlagAxelarChainName)
+			name, chainID, err := GetChainInfoFromFlags(cmd)
 			if err != nil {
 				return err
 			}
-			chainID, err := cmd.Flags().GetUint64(FlagAxelarChainID)
-			if err != nil {
-				return err
-			}
-
-			if name != "" {
-				req.ChainName = name
-			}
-			if chainID != 0 {
-				req.ChainId = chainID
-			}
+			req.ChainName = name
+			req.ChainId = chainID
 
 			res, err := queryClient.QueryScheduledCorks(cmd.Context(), req)
 			if err != nil {
@@ -190,21 +161,12 @@ func queryScheduledCorksByBlockHeight() *cobra.Command {
 				BlockHeight: uint64(height),
 			}
 
-			name, err := cmd.Flags().GetString(FlagAxelarChainName)
+			name, chainID, err := GetChainInfoFromFlags(cmd)
 			if err != nil {
 				return err
 			}
-			chainID, err := cmd.Flags().GetUint64(FlagAxelarChainID)
-			if err != nil {
-				return err
-			}
-
-			if name != "" {
-				req.ChainName = name
-			}
-			if chainID != 0 {
-				req.ChainId = chainID
-			}
+			req.ChainName = name
+			req.ChainId = chainID
 
 			res, err := queryClient.QueryScheduledCorksByBlockHeight(cmd.Context(), req)
 			if err != nil {
@@ -236,21 +198,12 @@ func queryScheduledBlockHeights() *cobra.Command {
 			queryClient := types.NewQueryClient(ctx)
 			req := &types.QueryScheduledBlockHeightsRequest{}
 
-			name, err := cmd.Flags().GetString(FlagAxelarChainName)
+			name, chainID, err := GetChainInfoFromFlags(cmd)
 			if err != nil {
 				return err
 			}
-			chainID, err := cmd.Flags().GetUint64(FlagAxelarChainID)
-			if err != nil {
-				return err
-			}
-
-			if name != "" {
-				req.ChainName = name
-			}
-			if chainID != 0 {
-				req.ChainId = chainID
-			}
+			req.ChainName = name
+			req.ChainId = chainID
 
 			res, err := queryClient.QueryScheduledBlockHeights(cmd.Context(), req)
 			if err != nil {
@@ -291,21 +244,12 @@ func queryScheduledCorksByID() *cobra.Command {
 				Id: id,
 			}
 
-			name, err := cmd.Flags().GetString(FlagAxelarChainName)
+			name, chainID, err := GetChainInfoFromFlags(cmd)
 			if err != nil {
 				return err
 			}
-			chainID, err := cmd.Flags().GetUint64(FlagAxelarChainID)
-			if err != nil {
-				return err
-			}
-
-			if name != "" {
-				req.ChainName = name
-			}
-			if chainID != 0 {
-				req.ChainId = chainID
-			}
+			req.ChainName = name
+			req.ChainId = chainID
 
 			res, err := queryClient.QueryScheduledCorksByID(cmd.Context(), req)
 			if err != nil {
@@ -345,21 +289,12 @@ func queryCorkResult() *cobra.Command {
 				Id: corkID,
 			}
 
-			name, err := cmd.Flags().GetString(FlagAxelarChainName)
+			name, chainID, err := GetChainInfoFromFlags(cmd)
 			if err != nil {
 				return err
 			}
-			chainID, err := cmd.Flags().GetUint64(FlagAxelarChainID)
-			if err != nil {
-				return err
-			}
-
-			if name != "" {
-				req.ChainName = name
-			}
-			if chainID != 0 {
-				req.ChainId = chainID
-			}
+			req.ChainName = name
+			req.ChainId = chainID
 
 			res, err := queryClient.QueryCorkResult(cmd.Context(), req)
 			if err != nil {
@@ -391,21 +326,12 @@ func queryCorkResults() *cobra.Command {
 			queryClient := types.NewQueryClient(ctx)
 			req := &types.QueryCorkResultsRequest{}
 
-			name, err := cmd.Flags().GetString(FlagAxelarChainName)
+			name, chainID, err := GetChainInfoFromFlags(cmd)
 			if err != nil {
 				return err
 			}
-			chainID, err := cmd.Flags().GetUint64(FlagAxelarChainID)
-			if err != nil {
-				return err
-			}
-
-			if name != "" {
-				req.ChainName = name
-			}
-			if chainID != 0 {
-				req.ChainId = chainID
-			}
+			req.ChainName = name
+			req.ChainId = chainID
 
 			res, err := queryClient.QueryCorkResults(cmd.Context(), req)
 			if err != nil {
